+++ source/MRVoxels/MRVoxels.vcxproj
@@ -44,6 +44,8 @@
     <ClCompile Include="MRTetrisNesting.cpp" />
     <ClCompile Include="MRToolPath.cpp" />
     <ClCompile Include="MRVDBConversions.cpp" />
+    <ClCompile Include="MRVDBConversionsToVolume.cpp" />
+    <ClCompile Include="MRVDBConversionsToSimple.cpp" />
     <ClCompile Include="MRVolumeSegment.cpp" />
     <ClCompile Include="MRVoxelGraphCut.cpp" />
     <ClCompile Include="MRVoxelPath.cpp" />
